Add default theme colors. Fix adding events. More precise number for progress indicator in detail view.

+++ Time Left/SKProgressIndicator.h
@@ -10,7 +10,7 @@
 
 @interface SKProgressIndicator : UIView
 
-@property (assign, nonatomic) NSInteger percentInnerCircle;
+@property (assign, nonatomic) CGFloat percentInnerCircle;
 @property (assign, nonatomic) CGFloat percentOuterCircle;
 
 @property (strong, nonatomic) UIColor *innerCircleBackgroundColor;
